requirements.txt
fastapi==0.65.2
gunicorn==20.1.0
uvicorn==0.14.0
numpy==1.19.5
Pillow==8.3.0
python-multipart==0.0.5
pytest==6.2.4
requests==2.26.0
aiohttp==3.7.4.post0
opencv-python-headless==4.5.5.62
imutils==0.5.4